Setup scripts and stuff!
Create one update/setup script for all -nix Like OSes, versus doing special stuff per OS.

+++ script/setup_linux.sh
@@ -1,10 +1,31 @@
 #!/bin/bash
+
+PUBPROFILE="Linux";
+if [ -f "/usr/bin/raspi-config" ] 
+  then
+    PUBPROFILE="LinuxARM"
+fi
+
+unameOut="$(uname -s)"
+if [ $unameOut = "Darwin" ]
+  then
+    PUBPROFILE="OSX"
+fi
+
+if [ $unameOut = "FreeBSD" ]
+  then
+    PUBPROFILE="Portable"
+fi
+
+echo "Beginning Glimmr setup/update for $PUBPROFILE"
+
 # Add user if not exist
 id -u glimmrtv &>/dev/null || useradd -m glimmrtv
 usermod -aG sudo glimmrtv 
 cd /home/glimmrtv || exit
+
 # Check dotnet installation
-if [ ! -d "/opt/dotnet" ]
+if [ ! -f "/opt/dotnet/dotnet" ]
 then 
   echo "Installing dotnet."
   echo "Downloading..."
@@ -13,15 +34,60 @@ then
 fi
 
 cd /opt/dotnet || exit
-# Install dependencies
-echo "Installing dependencies..."
-# Add extra repos for users on Ubuntu 20- for libjasper-dev
-sudo add-apt-repository "deb http://security.ubuntu.com/ubuntu xenial-security main"
-# Add extra repos for users on Ubuntu 20- for libqtgui4 libqt4-test
-sudo add-apt-repository ppa:rock-core/qt4 -y
-sudo apt-get -y update && apt-get -y upgrade
-sudo apt-get -y install libgtk-3-dev libhdf5-dev libatlas-base-dev libjasper-dev libqtgui4 libqt4-test libglu1-mesa libdc1394-22 libtesseract-dev scons icu-devtools libjpeg-dev libpng-dev libtiff-dev libavcodec-dev libavformat-dev libswscale-dev libv4l-dev libxvidcore-dev libatlas-base-dev gfortran libopengl-dev git gcc xauth avahi-daemon x11-xserver-utils libopencv-dev python3-opencv unzip
-echo "DONE!"
+
+if [ ! -f "/home/glimmrtv/firstrun" ]
+then
+  echo "Starting first-run setup..."
+  if [ $PUBPROFILE = "LinuxARM" ] 
+  then
+    echo "Installing Linux-ARM dependencies..."
+    # Install dependencies
+    sudo apt-get -y update && apt-get -y upgrade
+    sudo apt-get -y install libgtk-3-dev libhdf5-dev libatlas-base-dev libjasper-dev libqtgui4 libqt4-test libglu1-mesa libdc1394-22 libtesseract-dev scons icu-devtools libjpeg-dev libpng-dev libtiff-dev libavcodec-dev libavformat-dev libswscale-dev libv4l-dev libxvidcore-dev libatlas-base-dev gfortran libopengl-dev git gcc xauth avahi-daemon x11-xserver-utils
+    
+    #Assign existing hostname to $hostn
+    hostn=$(cat /etc/hostname)
+    hostn=`cat /etc/hostname | tr -d " \t\n\r"`
+    serial=$(cat /proc/cpuinfo | grep -E "^Serial" | sed "s/.*: 0*//")
+    newhost=glimmr-"${serial: -4}"
+    if [ "$hostn" != $newhost ]; then
+      echo "Changing hostname from $hostn to $newhost..." >&2
+      echo $newhost > /etc/hostname
+      sudo sed -i "s/127.0.1.1.*$CURRENT_HOSTNAME\$/127.0.1.1\t$NEW_HOSTNAME/g" /etc/hosts
+    fi
+    
+    if [ ! -d "/home/glimmrtv/ws281x" ]
+    then
+      echo "Cloning LED libraries..."	
+      git clone https://github.com/jgarff/rpi_ws281x /home/glimmrtv/ws281x
+      cd /home/glimmrtv/ws281x || exit
+      echo "LED setup..."
+      scons
+      gcc -shared -o ws2811.so *.o
+      cp ./ws2811.so /usr/lib/ws2811.so
+      echo "DONE!"
+    fi
+    echo "Raspi first-config is done!"
+  fi
+
+  if [ $PUBPROFILE = "Linux" ] 
+  then
+      echo "Installing Linux (x64) dependencies..."
+      # Add extra repos for users on Ubuntu 20- for libjasper-dev
+      sudo add-apt-repository "deb http://security.ubuntu.com/ubuntu xenial-security main"
+      # Add extra repos for users on Ubuntu 20- for libqtgui4 libqt4-test
+      sudo add-apt-repository ppa:rock-core/qt4 -y
+      sudo apt-get -y update && apt-get -y upgrade
+      sudo apt-get -y install libgtk-3-dev libhdf5-dev libatlas-base-dev libjasper-dev libqtgui4 libqt4-test libglu1-mesa libdc1394-22 libtesseract-dev scons icu-devtools libjpeg-dev libpng-dev libtiff-dev libavcodec-dev libavformat-dev libswscale-dev libv4l-dev libxvidcore-dev libatlas-base-dev gfortran libopengl-dev git gcc xauth avahi-daemon x11-xserver-utils libopencv-dev python3-opencv unzip
+      echo "DONE!"
+  fi
+  if [$PUBPROFILE = "OSX"]
+  then
+    echo "Hey there, I see you're trying to run this on OSX. You should leave a comment in the github issues section so we can fill this out!"
+  fi
+  echo "done" > "/home/glimmrtv/firstrun"
+fi
+
 
 if [ ! -d "/home/glimmrtv/glimmr" ]
 then
@@ -51,24 +117,40 @@ fi
 
 # Build latest version
 echo "Building glimmr..."
-/opt/dotnet/dotnet publish /home/glimmrtv/glimmr/src/Glimmr.csproj /p:PublishProfile=Linux -o /home/glimmrtv/glimmr/bin/
+/opt/dotnet/dotnet publish /home/glimmrtv/glimmr/src/Glimmr.csproj /p:PublishProfile=$PUBPROFILE -o /home/glimmrtv/glimmr/bin/
 echo "DONE."
+chmod -R 777 /home/glimmrtv/glimmr/bin
 
 #Get libcvextern.so
-sudo mkdir /home/glimmrtv/glimmr/lib/linux/libcvextern
-cd /home/glimmrtv/glimmr/lib/linux/libcvextern || exit
-sudo wget https://www.nuget.org/api/v2/package/Emgu.CV.runtime.ubuntu.20.04-x64/4.5.1.4349
-unzip 4.5.1.4349
-sudo cp /home/glimmrtv/glimmr/lib/linux/libcvextern/runtimes/ubuntu.20.04-x64/native/libcvextern.so /home/glimmrtv/glimmr/lib/linux/libcvextern.so
-sudo rm -r /home/glimmrtv/glimmr/lib/linux/libcvextern
+if [ $PUBPROFILE = "Linux"]
+then
+  echo "Installing libs for emgu.cv (Linux x64)"
+  sudo mkdir /home/glimmrtv/glimmr/lib/linux/libcvextern
+  cd /home/glimmrtv/glimmr/lib/linux/libcvextern || exit
+  sudo wget https://www.nuget.org/api/v2/package/Emgu.CV.runtime.ubuntu.20.04-x64/4.5.1.4349
+  unzip 4.5.1.4349
+  sudo cp /home/glimmrtv/glimmr/lib/linux/libcvextern/runtimes/ubuntu.20.04-x64/native/libcvextern.so /home/glimmrtv/glimmr/lib/LinuxARM/libcvextern.so
+  sudo rm -r /home/glimmrtv/glimmr/lib/linux/libcvextern
+fi
+
+if [ $PUBPROFILE = "LinuxARM"]
+then
+  echo "Installing libs for emgu.cv (Linux x64)"
+  sudo mkdir /home/glimmrtv/glimmr/lib/linux/libcvextern
+  cd /home/glimmrtv/glimmr/lib/linux/libcvextern || exit
+  sudo wget https://www.nuget.org/api/v2/package/Emgu.CV.runtime.linux-arm/4.5.1.4349
+  unzip 4.5.1.4349
+  sudo cp /home/glimmrtv/glimmr/lib/linux/libcvextern/runtimes/linux-arm/native/libcvextern.so /home/glimmrtv/glimmr/lib/Linux/libcvextern.so
+  sudo rm -r /home/glimmrtv/glimmr/lib/linux/libcvextern
+fi
 
 # Copy necessary libraries
 echo "Copying libs..."
 cp -r /home/glimmrtv/glimmr/lib/bass.dll /usr/lib/bass.dll
-cp -r /home/glimmrtv/glimmr/lib/linux/* /usr/lib
-
-cp -r /home/glimmrtv/glimmr/src/ambientScenes /bin/ambientScenes
-cp -r /home/glimmrtv/glimmr/src/audioScenes /bin/audioScenes
+if [ -d "/home/glimmrtv/glimmr/lib/$PUBPROFILE" ]
+then
+  cp -r /home/glimmrtv/glimmr/lib/$PUBPROFILE/* /usr/lib
+fi
 
 # Check service start/install
 if systemctl --all --type service | grep -q "$serviceName";then
